SHRL-53 refactoring
