+++ src/main/java/com/github/orionlibs/spring_http_request_logger/config/ConfigurationService.java
@@ -1,4 +1,4 @@
-package com.github.orionlibs.spring_http_request_logger;
+package com.github.orionlibs.spring_http_request_logger.config;
 
 public class ConfigurationService
 {
