+++ src/test/java/com/github/orionlibs/spring_http_request_logger/log/ListLogHandler.java
@@ -1,4 +1,4 @@
-package com.github.orionlibs.spring_http_request_logger.config;
+package com.github.orionlibs.spring_http_request_logger.log;
 
 import java.util.ArrayList;
 import java.util.List;
